board API에서 postBoard function 분리
