+++ src/api/boards/boards.ctrl.js
@@ -12,6 +12,8 @@ const {
     boardListScheme,
 } = require('../../lib/schemes');
 
+const { postFunction } = require('../../lib/functions');
+
 const {
     checkUser,
     checkBoard,
@@ -147,37 +149,7 @@ const getBoard = async (req, res, next) => {
  */
 const postBoard = async (req, res, next) => {
     try {
-        const { userId } = req.query;
-
-        const { error, value } = boardScheme.validate(req.body);
-        if (error) throw new Error('INVALID_PARAMETERS');
-
-        const { title, body, level } = value;
-
-        const user = await checkUser(userId);
-        if (!user) throw new Error('INVALID_PARAMETERS');
-
-        const { userLevel } = user;
-
-        const writeAuth = level <= userLevel;
-        if (!writeAuth) throw new Error('NO_AUTH');
-
-        const excerpt = body.join(' ').substring(0, 150);
-
-        /* ISSUE: 에러 발생하는 경우에 boardId 증가하지 않도록 (빈 번호 없도록) 처리 필요 */
-        await boards.create({
-            title,
-            body,
-            excerpt,
-            type: 'board',
-            hit: 0,
-            commentCount: 0,
-            userUserId: userId,
-            level,
-        });
-
-        /* TODO: 이미지, 파일 정보 추가 */
-        res.json({});
+        postFunction(req, res, next, 'board');
     } catch (err) {
         next(err);
     }
